+++ packages/tricorder-types/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to the [PVP](https://pvp.haskell.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- Require `effectful-core >=2.7 && <2.8`.
+- Require `effectful-plugin >=2.2 && <2.3`.
+
 ## [0.1.0.0] - 2026-08-24
 
 ### Added